+++ src/main/java/ai/aliz/jarvis/util/Helper.java
@@ -1,4 +1,4 @@
-package ai.aliz.talendtestrunner.helper;
+package ai.aliz.jarvis.util;
 
 import lombok.experimental.UtilityClass;
