+++ src/main/java/ai/aliz/talendtestrunner/service/SftpService.java
@@ -24,7 +24,7 @@ import org.springframework.stereotype.Service;
 import ai.aliz.jarvis.context.JarvisContext;
 import ai.aliz.jarvis.config.InitActionConfig;
 
-import static ai.aliz.talendtestrunner.helper.Helper.SOURCE_PATH;
+import static ai.aliz.jarvis.util.Helper.SOURCE_PATH;
 
 
 @Service
